+++ src/index.ts
@@ -1,1 +1,3 @@
+export * from './dashboard';
+export * from './datasource';
 export * from './grafana';
